+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## Changelog
 
+### [0.27.0](https://kaos.sh/pachca/0.27.0)
+
+- Added fields for audio and voice messages
+- Added field `CreatedAt` to `WebhookEvent`
+- Added method `AddBot`
+- Added method `GetBot`
+- Added method `EditBot`
+- Added method `RecreateBotToken`
+- Added method `RotateBotToken`
+- Added field `Sort` to define result sort order for message search
+
 ### [0.26.1](https://kaos.sh/pachca/0.26.1)
 
 - Added field `Chats` to `UserRequest`